Make es field configurable in transferToES.py
This allows the field we export to elasticsearch to be configurable. I'm
not sure if long term this will be our best option, most likely if we
decide to ship multiple scores with different names we want them to
go all at the same time rather than one per script run. This should be
a good enough starting point though.

Change-Id: I07e54ba3f7d412f941348a455965a7e8d3f23e7d