Fix errors found when deploying transferToES.py
Similar to the popularity score script, I found a few errors
that prevented deploying the script. This makes everything just
a little bit happier.

number of executors was also changed from 3 to 1. When i kicked
it off with 3 executors we jumped to ~1500 index/s and latencies
almost doubled. Dropping to a single executor will take longer
but seems a better bet in terms of maintaining reasonable
query latencies.

Change-Id: Ia9e6de7a4b09a0ebde92ba8c074f236b33894bac