Rework transferToES.py to send to wikis sequentially
The prior implementation of transferToES.py would send to whichever wiki
it had enough items to send to. This works but pushes the elasticsearch
cluster pretty hard, causing it to create segments in many different
indices at the same time.

This version reworks to use sort(project) which guarantees that each
project will be wholly contained within a single partition. It then
iterates that segment sending batches. This ensures that each worker
is continuously sending documents to the same index and reduces the
spread of segment creation, reducing load on the ES cluster.

The one downside of this is that foreachPartition() brings the entire
partition into memory, so the executors have to be sized such that any
executor can hold the entirety of whichever wiki has the most scores
to be sent. In the scheme of things this means using 15G of memory
for the job, which is 1% of total hadoop memory. It seems small enough
to be worthwhile.

Change-Id: Id49b74f5ce56ac8c845fd23798500d0476946500