Increase materialization timeout for transfer_to_es
It seems we can expect to wait 3 to 4 hours for the initial
oozie job to submit and complete. This import takes so long
it will run during the busy part of the day anyways. As a
once a week job it seems ok to allow this to wait up to 24
hours for the full data inputs before giving up.

Change-Id: I1c6655ffa580ff787382ad92f217fbcec761182f