Don't allow null values in popularity score
It turns out there are a couple nulls page_id's in here, which turns
into a failed request for every wiki when we transfer the data over to
es.

Additionally did a mild refactor to allow writing a test case. To run
the tests you need a full install of spark, simplest was to copy the
folder to stat1002 and run the test there through spark-submit.

Change-Id: I1856e9cd7434b066e5ad2dfa4174c89263a0a9cb