Repoint transfer_to_es job at specific servers
The firewall hole that was opened up between analytics and
elasticsearch does not include the LVS boxes. To work around
this point the job at a specific server in each cluster. It
might be better if we provided a list of servers, incase one
is down, but for now this is probably good enough.

Change-Id: Ic212cb1461de5ac9c1f012bed2f58490aee42d07