Rename 'detectors' to 'handlers'
For the super_detect_noop script to noop document updates
within 20% of the original value, it needs to read the configuration
for the handlers. This key was renamed in the search-extra plugin
in I3e032ca3 but the change never made it's way to CirrusSearch. As
such that mistake was propogated into this repository. Rename
detectors to handlers so the script starts nooping as expected.

This should, in theory, have a noticable impact on the # of documents
written out when the weekly dump hits.

Change-Id: I102031dfc96c3c14c465108e7786839f93f4ac68