Convert popularity_score into hql script
We were doing this calculation in spark because there was some
expectation that we would need to do something more complicated than the
simple % of pageviews. We've since decided this is probably fine as is.

There is a problem though, for some reason this is generating scores for
some pages as high as 0.98, which is plainly incorrect. Rather than
figure out what is wrong with the spark part, simplify it into a
straight forward hql that does everything we need.

Bug: T148136
Change-Id: I7f2ebc5b1a3d0d0edcbd9549c7cbb07abf17b3ae