transfer_to_es kafka tweaks
Talked it over and decided `cirrussearch.page-index-update` is
a better topic name, with the gola of being descriptive of what
the data inside the topic represents.

Also It's probably not a big deal, but if we are going to pull the job
down to a single executor we don't really want to force a giant sort.
Rather than sort take the list of projects and produce to them one at
a time. This also ensures if we allow producer parallelizm they will
be producing records for the same wiki.

Since the message volume here is high (benchmarked at ~12k messages/s
per producer vs mean of 2.5k/s for the eqiad-main cluster) very basic
rate limiting was added. The initial value is set to 1400/s which will
spread the data out over ~24 hours.

Bug: T200215
Change-Id: I2066c67905a1e679593d0e25ae6ed2a5e071914e