Deploy transfer_to_es with python dependencies
While testing this i found that the python kafka library is only
available on the stat100? machines, and not on the hadoop nodes so the
script doesn't work when executed on the cluster.

To resolve that port over the same deployment process used for
mjolnir. I'm ok with making a second version of some of the mjolnir
deployment stuff, but if there is ever a third we will want to
consider how to centralize.

Change-Id: I665af9d9de064902114e7f1031fd544569ca5f4f