Make transfer_to_es work with cdh5.15
* Don't use -- in filename

Since the cluster update to cdh5.15 (from 5.10) the transfer of
analytics data to the elasticsearch clusters has been broken. Looking
into it at some point the -- gets split from the argument, with spark
thinking it's a separate argument. Attempts to quote the paths made
no difference.

* add SPARK_HOME to transfer_to_es env

We removed this when upgrading from pyspark 1.6 to 2.3.1, as it was
correctly detected. With the cdh5.15 upgrade it seems SPARK_HOME is
back to undetected, so explicitly report where it is.

Bug: T210545
Change-Id: I391fb30b4f9625c0221d21f361bae64d6e43c670