Stop dropping hourly query clicks partitions
The daily rollups of the hourly data drop all the queries
that didn't result in clicks. The hourly data thus contains
a wealth of query reformulation information not retained in
the daily rollup.

There already exists a cron job that purges data older than 90
days for this table to comply with the privacy policy, so that
doesn't need to be added.

Change-Id: Iabe151c31bf7e913d3fc3c46ccb49b163d1af421