query_clicks: regression with newly added columns
The new request_set_token was added in an undeployable manner.
Alter table adds columns to the end of the schema, and insert
into table statements need to match their select to the tables
schema order.

Additionally during deployment hive decided that the partition keys
were ints. Not entirely sure, but plausible that the alter table
statement upgraded the schema to a new version that correctly
returns the partition keys as integers? Either way, cast to string
so concat_ws is happy.

Change-Id: I88047e895f9ea677000b41417b5e61d0f654e388