Remove cirrus request log dependency from query_clicks_daily
This was running an aggregation on the cirrus request log to get a
count of requests per ip per day. Rather than count on the bare log,
replace the count with one on the input hourly clicks log. This isn't
the same value, but probably a similar enough replacement. On average
the new numbers will certainly be lower, especially at the extreme
end.  On a random day, we had mean 8 req/ip and stddev of 400 in the
hourly click logs. The cirrus log for same day had a mean of 15 and a
stddev of 4000.

[1] mediawiki-event-schema/mediawiki/cirrussearch/request/0.0.1.yaml

Bug: T222268
Change-Id: I89ead7179396526645a4f9390cd1c200fcaa9d88