Move oozie hive actions to hive2
The hive actions are based on a old Hive protocol that
access directly HDFS and the Metastore. The Hive2 protocol
(used by Beeline) connects to the Hive Server 2 daemon via
JDBC and it is compatible with Kerberos authentication.

This is a first step to move towards hive2 actions, to verify
that they work properly before introducing the credentials/auth
configs.

Bug: T227257
Change-Id: I05f43d927dd6359b39ae5dae3925f5e75f7902a6