Import dym search satisfaction to druid
Transforms the TestSearchSatisfaction2 eventlogging schema into an
event per search and imports it into druid on a daily basis. At this
time this is only taking into account the did you mean functionality,
but it is expected we can extend this to handle other per-search
metrics.

The druid schema could probably be iterated on some more. It works for
the current purpose but might not be following druid best practices.
It is probable that with a better schema we could directly translate
source events into output events without directly doing per-search
aggregation, but would take further research and testing to implement.

Bug: T229268
Change-Id: Ie065938dd0bcf894c0d60cf2dc58983221e88fbd