Rename glent partition spec part -> date
part was an incredibly unspecific name of the partitions, but i
suppose allowed lots of flexibility. We already name the partitions
in YYYYMMDD format. Additionally refinery-drop-partitions script only
works with partition columns names it's familiar with, like date.
For these reasons rename part -> date.

Noticed this repo was missing the hql to create the glent tables as
well, those have been added.

Change-Id: Id354a727c4abada0920c0ed708076e72e6463eff