Define output paths more explicitly
We want to define the paths to use based off of hive metadata, to
avoid duplicating information that is stored elsewhere. We don't
want to query hive every time a DAG is built though since airflow
does that several times a minute.

We accomplished this previously with an _output_path method on
MjolnirOperator, and pushing the output path into xcom when a task
completes. This worked, but made one-off tests difficult as there
were no values to pull from xcom.

To accomplish the same goals, define _output_path as a templated
string and implement a jinja2 filter from the DAG that transforms
a table name into it's path. This allows sharing the templated
string between operators and completely avoids needing xcom.

Change-Id: I8dc7c108ccd6ef2acd47779831c5861dde4f9644