Split SkeinPlugin into hook/operator
Sharing scripts that use SkeinOperator between dags is a bit
difficult. Conver the skein plugin to work like the spark-submit
integration, with implementation in the hook and the operator for
tieing that functionality into the dag. This allows making a more
specific SwiftUploadOperator which can be shared between dags.

Additionally push some path configuration into the SwiftUploadOperator
to simplify sharing with minimal configuration required.

Change-Id: Ie2b77272418a7b9e153747a424da2d8f49d2b430