Rename drafttopics to articletopics
For the moment the input table still has drafttopic in the
predictions, but this will change to articletopic "soon". Transition
the elasticsearch ingestion pipeline to source from the old name
while emitting everything to the new name. Once the predictions are
renamed we should only need to adjust the python script CLI args.

Bug: T240556
Change-Id: Icccdfe4ea6076d6b46c90a138193fc9000d1cc66