skein_operator: Ensure hook is creatable
Only the SkeinHook is currently used, so this didn't cause any
problems, but the Operator was unable to create the hook due
to mixups in which properties are prefixed with a leading _. Add
a test for this bit of code and correct the operator. Idealy this
would have been caught by mypy, but airflow doesn't export types and
we don't have any stubs.

Change-Id: If9431221095ea5731bc0d2edaf4c26811385f8af