Dont directly refer to runtime paths in operators
Detecting and injecting the REPO_BASE into operators means that test
case outputs vary depending on where in the filesystem the repository
is checked out to. Avoid this by adding a new variable to point at
the path, and let templating resolve that when running the operator.
The prod installation will also need this variable set.

This additionally exposed that changing the REPO_BASE didn't change
any fixtures, so generic fixtures for SparkSubmitOperator have been
added to capture changes to those jobs and make sure we understand
when they are being changed.

Change-Id: I437b0a00742890c47bba955366b241ec67187e51