Rework virtualenv handling for airflow
We previously built virtualenv building into this repository to
support providing kafka dependencies for the transfer_to_es script.
That now works via eventgate so it was unused. Rework to be less
specific to oozie by having an environments directory, which contains
sub directories that define individual python exection environments.
Adding a new directory with a requirements.txt file will generate
a matching virtualenv on deployment.

* Drop old kafka artifacts, they are unreferenced.
* Look for requirements.txt in environments/* instead of oozie/*
* Remove debian jessie specific upgrades of pip and wheel
* Move virtualenv script from a scap command to a script, scripts
 were added specifically for the use case but the repo was never
 updated.
* Force python to always use 3.7, this is the only version available
 on debian buster based instances (like airflow).
* Install frozen requirements before regular from make_wheels.sh, to
 prefer only upgrading packages when explicitly asked to.

Change-Id: I30d0d7b515f782c2dfa2e051f0fe16b9065117c7