Allow SparkSubmitOperator to run custom virtualenvs
Due to how the spark-env.sh on wmf hadoop is written we
need to be able to provide environment variables to the
spark-submit invocation. The upstream SparkSubmitHook
allows for this, MjolnirOperator uses it, but the SparkSubmitOperator
does not. Replace with our own implementation, as the upstream
version is a very simple wrapper around the hook.

While here write a test case that verifies spark is invoked
in a way the correctly runs our custom environments.

Change-Id: Iedef794a478daf33ffc32bef92d9bb7292ae0008