Build and deploy virtualenvs from debian stretch
Virtualenv's built on stretch work on buster, but not the other way
around. The debian buster based virtualenv's hit an issue with buster
using a newer glibc than stretch. Deal with this by building
virtualenvs on a stretch machine and shipping them to hdfs to be
referenced by the airflow operators.

Change-Id: I7384bfa95af1cbf4828fecb118c8336c02ca374b