Port search satisfaction data pipeline to airflow
The search satisfaction jobs in oozie failed and didn't bother sending
any failure emails. This is likely a misconfiguration on our end, but
rather than digging into the configuration of something we are moving
away from move the search satisfaction code from oozie to airflow.

This includes a reimplementation of the oozie workflow used to load
data into druid as HdfsToDruidOperator. This accepts the same
formatted spec's as the oozie workflow and applies the same
templating.

Change-Id: Ia635a761137d4a8811b4e477aecec6d0c86c80ed