Reuse test code for finding tasks
Per the pytest docs the use of `import conftest` from a non-package
(such as our tests here) can be ambiguous, so we've previously avoided
performing the import. Avoiding the import is causing duplication, and
on consideration it seems having reusable test functions is simply
more valuable than avoiding that ambiguity.

Change-Id: Id39d0d52da75c0345338d42c3a43edd76d0824a0