Add search result/suggestion provider to search satisfaction
To be able to generate metrics about the different query suggestion
providers we need to track who that provider was in our downstream
metrics. This data has been added to SearchSatisfaction logging
in Ib8ea403e. Propagate that down the line so it can be faceted on
in druid/superset.

This requires a migration of the discovery.search_satisfaction_daily
table to add the new columns.

Change-Id: Ic157265b88bcbd148c5747b1f6fd4316003240d7