Convert plugins to a python package
Letting airflow inject our custom hooks and operators into the airflow
module namespace as a plugin was problematic. One file couldn't refer
to code in another file, there was no reasonable way to share
information. Additionally mocking in tests was more difficult due to
being unable to access the plugin modules directly.

It turns out airflow adds the configured plugins directory to
sys.path, so the only real change necessary to access these classes
through typical python imports is to move them into a directory and
add __init__.py to make it a module. The remainder of this patch
is repointing imports and cleaning up effects of not previously
being importable.

Change-Id: I88ed6ef06e17b0fc97f9c4c44d60c5724e5efd0c