convert_to_esbulk: super_detect_noop handlers as first class citizens
Refactor convert_to_esbulk configuration to directly represent the
handlers so we can delegate implementation details to them.
Additionally implement functionality to allow multiple tables to
import to the same elasticsearch field if the chosen Field class
supports it. This delegation is required for a followup that adds
support for the new multilist functionality in super_detect_noop.

Change-Id: I8cdce8935702ed33fbbfc624f075d79f0dad92fc