Centralize reading/writing to hive
Introduces three new classes to the shared spark code to cover
the various use cases found in our current scripts. Two for reading,
and a third for writing. All are designed to be instantiated from a
string typically provided as a cli argument. The HivePartition class
provides the foundation and parses hive's string representation of
a partition into an object that can read the table. The other two
classes provide support for reading a time range of partitions, and
for writing to the partition.

The general idea here is to remove concerns about how data is
partitioned from the scripts that perform the transformations.  How
data is partitioned is a fairly cross-cutting concern and seems better
represented at the dag level where inputs and outputs are being
assembled. This simplifys standing up a new script, and adds
opportunity for expectations around dataframe schemas as we move
forward.

Due to the new abstraction script inputs are less strict. For example
a script that takes the same hour of two datasets no longer knows they
represent the same hour, it must trust that the inputs are correctly
aligned.

Change-Id: Ib7c65cafbec3fbac532a61ee13c1298041108780