airflow: Deploy variables via scap
Recently we've taken to use variables more liberally in the DAGs. This
has made the current manual deployment of variable changes
unmanagable. It turns out airflow has a straightforward process for
importing json files as variables, use that. Depends on puppet
change which gives the deployment user sudo rights to invoke
`airflow variables`.

Change-Id: I7eed511d105b74dbbca6a940b45e8a507860836f
Depends-On: I6f2fb4a14743b25eb5ff2fdb2c23f41d9074814c