convert_to_esbulk: Repair variable referenced before assignment
Not sure why this wasn't caught by flake8 or mypy, but it was
impossible for validate_config to return True. This function
was also lacking tests that invoke it and show that to be the case.

Add some test cases for the functionality that would have failed, and
apply a fix. While here add a few other conditions to validate_config
that should also fail, all related to `len(something) == 0`.

Change-Id: Ifc638550e449e4b3a121f287ae3538fa492b9c49