airflow: refinery_drop_hive_partitions: Fix kerberos token passing
While this ran successfully on the first deployment, it seems that was
because a user had manually used `kerberos-run-command` with the
analytics-search user, which puts credentials in the system default
location. The `airflow kerberos` command maintains it's credentials
separately. When passing `env` to BashOperator we no longer get a copy
of the host environment vars, which I think is excluding configuration
that tells hive where to find the kerberos credentials.

Work around by providing PYTHONPATH through /usr/bin/env, instead
of through the BashOperator `env` variable. Using /usr/bin/env will
correctly add to the envrionment rather than overwrite it. We could
alternately have copied os.environ and kept the whole thing, but that
seems quite messy.

Change-Id: I94b8bc32e10ebeb961f497bf807d0db0da15d1fe