Also fix column mismatch writing articletopic predictions
Hopefully the last in this series of aligning table partitioning
with reality. In the future we need some automated way that these
can be tested prior to deployment.

It's also a little awkward that the table is specified in a variable
but the date-based partitioning is hard coded. Seems they should be
held closer together but not sure how yet.

Change-Id: I4213b5850898ff13243ec814f8fdfcca83e7ff1a