hive/spark scripts for initializing and integration environment
Includes create table statements for tables that are created outside
analytics/refinery and our own repository for purposes of initializing
and integration environment with appropriate schemas.

Includes scripts to populate tables with small amounts of fake data.
These are very much WIP. The goal is to have a trivial amount of data
that can feed into airflow tasks we want to test.

Change-Id: I035f667dce779c2ba3f1dbb0781277814e88a846