airflow: Match airflow version in CI with prod
A recent bad release of airflow broke CI merges. Rather than letting
our tests upgrade the airflow code at will, match it to the deployment
versions to prevent regressions and ensure we are testing against what
we plan to deploy. Also explicitly uses Werkzeug 0.16.0, matching
prod, as tox was trying to install 1.0.1 which is incompatible.

Change-Id: I5a56f70c99aa83db464f6b5d596433fbedb01429