airflow: ores: Create path thresholds are written to
This is effectively a no-op in production, as this path was manually
created at some point in the past. Moving forward though it seems
more resilient to encode this kind of setup logic in the dag itself.
Using this convention everywhere will reduce the burden within the
integration environment to setup all the possible fiddly details of
individual dags by encoding the needs within the dag.

Change-Id: I92eb78bedb0b8cad53802f796faa90efcc07471c