Centralize partition_spec templates
We have many places that need to ask for the partition matching
the execution date of the task. Define the pattern in
wmf_plugin.template and reuse instead of having multiple definitions.

The simplification to use integer values of the `execution_date`
objects in macros instead of time formatting also means time is no
longer prefixed to constant width (e.g. day=1 instead of day=01).
At one time, in the oozie days with exact paths, this was important. I
double checked today and all our uses of these go through query
execution, and all of the referenced fields use either int or bigint to
represent the portitons of a datetime. This no longer matters.

Change-Id: I4226a958d329d7ec7b5b7bf0a395faa2f261be0d