Standardize access to variables
when referencing the variables we load from /airflow/config/*.json each
use case had it's own implementation. This was basically a single line
of formatting, not a big deal. In short text, such as a path, attaching
a few templated strings emit from this method is just fine. In longer
text, such as a templated HQL block, using two separate templating
systems makes things less obvious and more error prone.

Add a standardized implementation of accessing variables. This method
exposes both the existing functionality and a new set of functionality
to make the variables available in the templates with a shorthand syntax.

As this is encouraging additional usage of templates add a new test
that ensures we are using StrictUndefined on all dags. Without this
templating will replacing missing values with the empty string.

Change-Id: I68ceaa01b9e33f6477f95ac3bf3df06a8f308070