airflow: Properly align threshold paths across days
Followup to I3a49d6 and I3e416c. When changing the threshold paths to
refer to yesterday, since the daily task run at midnight is dated with
the previous day, I mistakenly changed both the path to write to, and
the path to read from. Changing the path to write to means we are trying
to also write paths to the previous day.

Make separate functions for the two different threshold_path references.

Change-Id: I5c0208bb8f500b10bbe124995d64d812b475e792