Migrate ores topic tables to hourly partitioning
Missed this step when putting together the hourly runs. The tables
must be created with hourly partitioning to match the run schedule.

The _init dags are run with schedule of @once, so we can't change
the existing dag. Rename the _init dag to _v2_init to hold the
next stage of initialization.

While renaming drop the statement to rename wbitem table, it's already
been migrated.

Change-Id: I1136a28ad42a2e57ec10ed54d8f77ae1c66e56b7