ores: Partition staging tables by data source
To load dumps from ores into elasticsearch without creating one off
tables we need to make space for them in the typical staging tables.
Add an additional partitioning key to ores tables, use the dag_id
as the source when writing, and limit convert_to_esbulk to only
reading from the expected dag.

Add a new configuration to convert_to_esbulk that duplicates
the traditional config, but only contains ores tables and
repoints the partition_spec_tmpl at the dumps.

When updating the rel_path_* variables, the versioning is being removed
from the directory names. In review of the last deployment of
partitioning changes hive moved the data from it's specified location to
the hive warehouse (hdfs:///user/hive/warehouse) when renaming the
table. Since the old data is moved out of the way there is no need to
rename the live location when updating the schema.

Change-Id: I49991a3c4e5c2353d4e2d370d8f5a0b41ae322c4