airflow: Add tests for valid hql syntax
It seems difficult to get these completely correct, even when
reviewing the final files it's easy to overlook an extra comma
or a missing single quote. Add spark to the airflow test suite
and use it's parser to validate the hql statements are valid.

This is a bit awkward. These queries are actually parsed by hive,
not spark, and the spark parser only works on single statements.
The kludges required seem worth the benefit of knowing these
are at least probably valid hql.

Also correct the ores_predictions_init_v3 dag, which couldn't run
on deployment due to invalid hql.

Change-Id: Id1f5dcd7adb576adfbe8a7be18fcea008434ceb4