Move query_clicks to airflow
This has been our last job living in oozie for some time, move it
over to the airflow side to simplify things. This still doesn't fit
with the more recent code, it's a pair of overly complicated HQL
queries instead of a spark job, but this is a step in the right
direction.

Solves a recurring problem with sourcing namespace data, in oozie
we used a constant reference to a specific partition of a table
to read, within airflow we instead use the `hive.max_partition`
macro to always reference the most recent namespace data.

Prior to merging the existing oozie coordinators need to be killed
and the start time's in these airflow coordinators need to be updated
to pick up where they left off.

Bug: T298648
Change-Id: I116f81cc292226b4e0d0c95b012d4ac0edfe45ce