airflow: Make transfer_to_es functionality usable anywhere
We have various dags that all want to do the same thing, run
convert_to_esbulk.py over some configuration and upload the result to
swift. Today it's only the two in transfer_to_es, but a third for
manually triggered ores imports is also in progress. Move the
implementation to a plugin function that returns operators that can be
attached to any dag.  While here did some minor cleanup of the other
transfer dags to try and make it clear how much is shared between them.

Change-Id: Id6a58715a307de24768b39dc536951380c297689