query_clicks: Remove result file merging
The recent upgrade to hive 2.x has caused this to start throwing
a NullPointerException. This appears to be the same as
https://issues.apache.org/jira/browse/HIVE-16958.

Fix by removing result file merging. In a test execution this resulted
in 14 35MB files instead of the more typical 4 128MB files. Having less
files was an optimization, and not required as long as the partitions
aren't particular small or generating huge numbers of partitions.

Change-Id: Ie1798764c37d74fee95264feb7a64529b797e863