import_wikidata_ttl: Simplify time handling with pendulum
When starting with airflow we used airflows ds_* macros for most
time handling, but overall these are verbose and somewhat difficult
to use. Airflow also uses pendulum datetimes which are designed
for readability, use those instead to make readable expresions.
We can be confident these new expressions are the same as the fixtures
of their render is unchanged.

Change-Id: I87c4fc842455c2996db1a7f9ac9552ca2a0cf931