Share rendered_task test fixture between test files
Rendering a task is generally useful thing that tests
want to do. We already had duplicates of this functionality
in two places. Move the fixture from test_dag_structure to
conftest where it can be shared and reuse in the skein
plugin.  Minor fixture change is due to the test date for
skein changing.

Change-Id: I9ff8f33a7a329fcd709a63450994e6ea60c23ed3