airflow fixtures: Use valid schedule date for fixture rendering
In I87c4fc84 some time handling was changed in an incorrect manner yet
the rendered fixtures stayed exactly the same. This came about because
the functionality was dealing with a specific day of the week, but the
fixtures were rendered against a day that was not valid according to the
schedule.

Change fixtures to always render against the first valid date. This
ensures we meet the expectations of the tasks being rendered. This does
mean pre-deploy changes to start_date will trigger a fixture rebuild,
but the benefit of catching an error we've previously run into seems
worthwhile.

Using the date in fixtures means all dags must have a constant
start_date. Changed only usage of a dynamic start date, which wasn't
doing anything useful, to a constant date. This is also documented
in the airflow faq as best practice.

Change-Id: I6c0843b32da4c60c84694abc54cafdafd0437354