Build virtualenvs inside the yarn cluster
We need some way to build these virtualenv's for deployment. They can no
longer be built on nodes we use to access the yarn cluster, they run a
different version of debian. To get things moving ship the repository
to a node inside the cluster and build the virtualenvs there. Push the
built env's up to hdfs to later be referenced.

This does have an issue, the path to read from is commited to this
repository. That means we must commit environment changes as separate
patches, deploy, run the script, and then ship a second patch that
updates airflow/config/wmf_conf.json with the correct path and uses
whatever functionality is in the new venv builds. If this was a long
term solution we could extend this to having scap build the envs and
update an airflow variable on each deploy, but analytics is looking to
make conda the standard package manager so we will likely be switching
away from this solution as they figure that out.

Change-Id: I71c60fa41598127735d347a44e9633c07a5f435d