convert_to_esbulk: Accept partial hour timestamps
This restriction was put in place as it was assumed all data coming in
would be scheduled directly on an hourly boundary. This is not actually
the case, manually triggered dags (such as ores bulk export) have
timestamps with no particular boundary. Relax the requirement to allow
the export to process.

Change-Id: I29ba8618a33b1e02cc32ae4259af3e6878fdef88