glent: Turn on version marker
We added a version marker which, in theory, allows to ask the frontend
with a specially formatted query if glent is enabled and which version
of the dataset is being provided. This will emit the marker for
testwiki, we will need to turn glent on for testwiki for this to be
findable.

Change-Id: I4f6af3ee4e16339b88939a78e1f8542b3ac608e0