Deploy mjolnir venv through environments
Up until now mjolnir has had it's own custom deployment scheme,
written well before this repo had a way to manage it. Now that we
have a standardized way of deploying python environments move this
over.

The wmf_mjolnir package is still not released to pypi or any other
proper package index. Rather the mjolnir repository has been updated to
support building a wheel which must be copied to the artifacts directory
here. After the wheel has been built it was copied into the artifacts
here as well. The make_wheels.sh script had some minor changes applied
to work with packages that are not released to pypi.

Bug: T265547
Change-Id: I410c1dcea90e6037cf247c0296813c849d56c816