Match python_logstash wheel to sha1 in archiva
Python packages aren't deterministic, so it's possible to have multiple
packages of the same version. Typically this is caught by running the
wheel upload process prior to uploading the git patch, but it seems this
was skipped last time. Ideally we need a more automated process, but for
now simply update the hash to match.

While here, the upload script didn't entirely work. It errored out
trying to access a python2 function while the shebang calls out
python3. Updated urlretrieve to use the correct py2/3 compat module.

Change-Id: Ieaa9f11f566a5782b4376f3b0850d1deca48dc91