Ship packages for pip and wheel with virtualenvs
Attempting to build the mjolnir venv is failing due to pip not
recognizing one of the packages and being installable. Testing verifies
that while the system pip version, 18, refuses to install, updating to
pip 21 allows the install to continue. This same functionality was
previously implemented in the mjolnir-deploy repo, this is bringing over
the solution to this repository.

Change-Id: I9e02c9a6eefda6d645639c3d5e2e5ea0179fc1c4