Point at correct glent 0.2.6 jar
The file uploaded to archiva and the file exposed by archiva seem to be
slightly different. The prior jar was taken from the target/ directory
after performing a release, but that doesn't seem to match the release
jar available from archiva.

Update reference here to use the hash archiva is reporting.

Change-Id: I91768c2f64be0d1a5db2a6e5938d8261bf67e985