Expect eventgate canary events in all dcs
Eventgate used to only generate events for the currently active
mediawiki datacenter. For some time (not sure when exactly) now
eventgate is producing canary events to both datacenters, meaning we
should be able to wait on all datacenters instead of just the
active one. This was previously requested by us (and others), and
removes the need to change the active datacenter when it moves.

Implements a new TemplatedVar object in wmf_airflow.template. This
helper accepts a macro expression that resolves to an iterable, formats
the iterable to a string for jinja2 output, then re-parses back into an
array of values on use. To support existing use cases this helper
accepts a transform function to apply to resulting list elements, and a
sequence of arguments that will also be rendered and passed to the
transform fn.

A bit of a hack, but all templated outputs must be strings. Additionally
we typically have templated strings we want to mix into this value.
Future work may want to consider accessing the Variable object directly,
but that seemed more complicated and had to deal with concerns around
when, exactly, the Variable value should be loaded from the database.
In airflow 2.1 the list->str->list can possibly be avoided by setting
render_template_as_native_obj=True.

Change-Id: I74db82f5e8c83696333c996876e348c5ebb770b7