export queries to relforge: Allow for null columns
We've seen intermittent failure of this job. Looking into a single
hour of failure what happened is the query syntax field was null for
a single row, and spark generated code to process those row's expecting
there were no nulls as was declared in the program. We don't actually
have strong guarantees on nullability, the hive datastore does not
include nullability constraints for accessed tables.

Change the nullability declaration to allow nulls, which in testing
allows the jobs to finish processing previously failing hours.

Change-Id: I056baa9b5d066b8e97311089669670b28fd2f4f5