airflow: Move hive_table_path into the plugin
This was a bit of a confusing implementation. This filter was
only defined in the mjolnir dag, but then used from the plugin.
It seems for consistency the filter should exist in the plugin
and only be referenced from the dag as necessary.

Additionally some new tests are being added in a followup that
need to mock out this filter to ensure it doesn't try to talk
to hive during unit testing. It's easier to mock out when it
can be imported directly.

Change-Id: Ifab447a469d190c4a9b8162655dd08983cb899ad