Increase subgraph_mapping_weekly executor memory
While monitoring the first run of this the 6th stage with 8
partitions repeatedly failed 2 of the partitions, eventually
failing the job.  Many of the executors are reporting as red
(>10% of execution time spent inside the GC) in the Executors
tab of the Spark UI.

Increase executor memory by 50% to get this running. If it
continues to be problematic we could test changing
`topSubgraphItems.coalesce(8)` in SubgraphMapper to
`topSubgraphItems.repartition(8)` which will likely
compute with 200 shards and then shuffle down to 8
after the computation is complete.

Change-Id: I8b344b82fc76cd440fcd90e2604a00d9b1fdd6f4