image_suggestion_manual should read from _delta not _full
Talked with the team providing these suggestions after they noticed we
didn't have the updated data they were expecting. It turns out the
problem is the manual import always referred to _full, only the
automated weekly import (which was disabled because no automated weekly
data is available) read from _delta. Change the manual import over
to the _delta as well.

Change-Id: Ic9e537bac1c5007c8936aa154c59d039782e36fb