Increase subgraph query SLAs to 2 days
These SLAs are a bit spammy and usually there is nothing in particular
to do about them, nothing went wrong and we need to wait a little longer
for them to complete. Increase the SLA to 2 days so we are only notified
when its more likely to need operator intervention to get things moving.

Change-Id: I305d37f02015bf0df57c0de5a524c3db599c1384